requirements.txt
Flask
tensorflow
Pillow
numpy
gunicorn